requirements.txt
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dialogflow==1.0.0
Flask==1.1.2
google-api-core==1.22.1
google-auth==1.20.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pipdeptree==1.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
twilio==6.45.0
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
